订票初步完成

+++ src/main/java/com/wyx/amovie/entity/Order.java
@@ -19,6 +19,6 @@ public class Order {
     private Date createTime;
     private Integer sceneId;
     private String ticketNum;
-    private Float totalPrice;
+    private Integer totalPrice;
     private String seat;
 }
